feat : React 패키지 구조 분리 & 백엔드 API 호출 & 로컬스토리지에 엑세스 토큰 가져오기
